python/requirements.txt
requests
flask
pymysql
psycopg2
flask-cors
bcrypt
pycrypto
pyjwt